msw: Convert root `handlers/` to TypeScript
This converts the remaining top-level handler modules (`cdn`, `docs-rs`,
`github`, `gitlab`, `metadata`, `native-replacements`, `playground`,
`rustsec`, and `summary`). `cdn.ts` types its catch-all path parameters
and archive helpers.
